requirements.txt
Flask==3.1.0
cryptography==44.0.2
psutil==5.9.8
mysql-connector-python==9.2.0
gunicorn==21.2.0
python-dotenv
pycryptodome==3.17.0